pyproject.toml
[project]
name = "learn-grpc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.35.0",
    "websockets>=15.0.1",
]
